Moved and renamed NoteService file to /api/note.js. Renamed NoteComponent.vue to Note.vue
